+++ UsersMicroservice/BusinessLogic/Services/UserTransactionService.cs
@@ -1,4 +1,5 @@
 ﻿using BusinessLogic.Services.Interfaces;
+using BusinessLogic.Validation.Validators.Interfaces;
 using DataAccess.Entities;
 using DataAccess.UnitOfWork.Interfaces;
 
@@ -6,24 +7,41 @@ namespace BusinessLogic.Services
 {
     internal class UserTransactionService : BaseService<UserTransaction>, IUserTransactionService
     {
-        public UserTransactionService(IUnitOfWork unitOfWork)
-            : base(unitOfWork)
+        public UserTransactionService(IUnitOfWork unitOfWork, IBaseValidator<UserTransaction> validator)
+            : base(unitOfWork, validator)
         {
         }
 
+        // Пока что я буду просто выбрасывать ошибки. Но в будущем будет выбрасывать DTO, хранящая ошибки,
+        // которые будет принимать контроллер и возвращать в frontend
         public override async Task CreateAsync(UserTransaction entity)
         {
-            await _unitOfWork.UserTransactionRepository.CreateAsync(entity);
-        }
+            var validationResult = await _validator.ValidateAndThrowAsync(entity);
+            // Это временно
+            if (validationResult.Any())
+                throw new Exception(string.Join(Environment.NewLine, validationResult.SelectMany(e => e.Value).ToList()));
 
-        public override async Task DeleteAsync(UserTransaction entity)
-        {
-            await _unitOfWork.UserTransactionRepository.DeleteAsync(entity);
+            await _unitOfWork.UserTransactionRepository.CreateAsync(entity);
         }
 
         public override async Task UpdateAsync(UserTransaction entity)
         {
+            var validationResult = await _validator.ValidateAndThrowAsync(entity);
+            // Это временно
+            if (validationResult.Any())
+                throw new Exception(string.Join(Environment.NewLine, validationResult.SelectMany(e => e.Value).ToList()));
+
             await _unitOfWork.UserTransactionRepository.UpdateAsync(entity);
+        }
+
+        public override async Task DeleteAsync(UserTransaction entity)
+        {
+            var validationResult = await _validator.ValidateAndThrowAsync(entity);
+            // Это временно
+            if (validationResult.Any())
+                throw new Exception(string.Join(Environment.NewLine, validationResult.SelectMany(e => e.Value).ToList()));
+
+            await _unitOfWork.UserTransactionRepository.DeleteAsync(entity);
         }
 
         public override IQueryable<UserTransaction> GetAll()
